requirements.txt
streamlit
torch
numpy
Pillow
albumentations==1.3.1
opencv-python-headless
reportlab